+++ package.json
@@ -33,5 +33,8 @@
   "engines": {
     "node": ">=4"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "@types/node": "^14.10.1"
+  }
 }
